requirements.txt
laion-clap
open_clip_torch
pillow
pytorchvideo
einops